requirements.txt
streamlit
tensorflow
numpy
Pillow # This is the package name for PIL
matplotlib